Simplify submitted payment_method_type validation
Instead of using `String#constantize`, which is scary, detect the string
from a known set of good classes. There were validation in place
previously to ensure admins could not constantize arbitrary classes, but
this reads safer, and would be less prone to a dangerous error in the
future.

Also replaced an `#update_columns` and `#reload` with a `#becomes`