Update outstanding_balance to use reimbursements
This updates outstanding_balance to use the value of "reimbursements"
when determining the payment total.

Previously the value of an order's refunds was used both for
representing the money returned to the customer as well as the reduced
"adjusted total" of the order (which we expected to be paid). This was
wrong in a number of cases.

This commit updates outstanding_balance to use the value of
reimbursements for calculating the "adjusted total" of the order. This
allows all cases to (at least) be represented properly. This better
matches the best practice double-entry bookkeeping style.

Reimbursements are created when going through the refund system so this
should continue to work for cases where refunds where made through the
returns system.