Fix failing payment spec
Update this spec to match the values it currently returns.

I changed the values in this spec into what values we got from the
system, which isn't ideal, but I think these values make sense.

We have a payment for an amount and a refund for the same amount, so
payment total should be zero. Because the order is cancelled, this is
the state we want, and the outstanding balance should be zero.