Add users to the quick switch component
Users is another resource that admins want to quickly access. This
particularly one has the potential to cause problems for stores that may
not implement `Spree.user_class` in the same way that
`solidus_auth_devise` does. For example, this component makes the
assumption that `email` exists on a user.