Remove keyboard navigation from display order page
This feature doesn't work well. It requires using "Tab" to get focus
on an item in the list. After doing that it allows pressing up/down to
reorder the list.

This feature isn't discoverable to users, so I doubt it has ever been
used. It also doesn't exist in any of our other sortable tables in the
admin.

It could just as be implemented using Sortable.js, which we are moving
to from jquery-ui's sortable, but I would not rather not spend the time
reimplementing something that doesn't work well and isn't consistent
with the rest of the admin. If we want this we need to implement is
across all of our sortable lists.