Remove overridden Admin::Variant#destroy action
This action has a funny history. It was originally created to set the
deleted_at column before the project used paranoia. At a later date,
when paranoia was included, this was changed to call destroy, since
paranoia causes that to just set deleted_at. At this time the override
probably could have been removed.

We've now gone almost full circle and are preferring discard over
paranoia. We don't need this method and can use the implementation in
the ResourceController.